// lib/api/generated/model/createDatingProfileRequest.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { CreateDatingProfileRequestCity } from './createDatingProfileRequestCity';
import type { CreateDatingProfileRequestDatingStatus } from './createDatingProfileRequestDatingStatus';
import type { CreateDatingProfileRequestInterestedGenderItem } from './createDatingProfileRequestInterestedGenderItem';
import type { CreateDatingProfileRequestInterestsItem } from './createDatingProfileRequestInterestsItem';
import type { CreateDatingProfileRequestReligion } from './createDatingProfileRequestReligion';
import type { CreateDatingProfileRequestReligiousPreference } from './createDatingProfileRequestReligiousPreference';

export interface CreateDatingProfileRequest {
  city: CreateDatingProfileRequestCity;
  bio?: string;
  /** @minimum 18 */
  ageFrom: number;
  /** @nullable */
  ageTo?: number | null;
  interestedGender: CreateDatingProfileRequestInterestedGenderItem[];
  religion: CreateDatingProfileRequestReligion;
  /** @nullable */
  religiousPreference?: CreateDatingProfileRequestReligiousPreference;
  interests: CreateDatingProfileRequestInterestsItem[];
  datingStatus?: CreateDatingProfileRequestDatingStatus;
}