// lib/api/generated/model/ownProfilePhoto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { PhotoSuggester } from './photoSuggester';

export interface OwnProfilePhoto {
  id: string;
  storageUrl: string;
  displayOrder: number;
  /** @nullable */
  approvedAt: string | null;
  /** @nullable */
  suggesterId: string | null;
  suggester: PhotoSuggester | null;
}