// lib/api/generated/model/ownProfilePrompt.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { OwnPromptResponse } from './ownPromptResponse';
import type { PromptTemplateRef } from './promptTemplateRef';

export interface OwnProfilePrompt {
  id: string;
  answer: string;
  createdAt: string;
  template: PromptTemplateRef;
  responses: OwnPromptResponse[];
}