// lib/api/generated/model/ownPromptResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { PromptResponseAuthor } from './promptResponseAuthor';

export interface OwnPromptResponse {
  id: string;
  message: string;
  isApproved: boolean;
  userId: string;
  createdAt: string;
  author: PromptResponseAuthor | null;
}