// lib/api/generated/model/profile.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { ProfileGender } from './profileGender';
import type { ProfileRole } from './profileRole';

export interface Profile {
  id: string;
  /** @nullable */
  chosenName: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  dateOfBirth: string | null;
  /** @nullable */
  gender: ProfileGender;
  role: ProfileRole;
  /** @nullable */
  pushToken: string | null;
}