// lib/api/generated/model/publicDatingProfile.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { PublicProfilePhoto } from './publicProfilePhoto';
import type { PublicProfilePrompt } from './publicProfilePrompt';

/**
 * @nullable
 */
export type PublicDatingProfile = {
  id: string;
  /** @nullable */
  bio: string | null;
  city: 'Boston' | 'New York';
  interests: (
    | 'Travel'
    | 'Fitness'
    | 'Cooking'
    | 'Music'
    | 'Art'
    | 'Movies'
    | 'Books'
    | 'Gaming'
    | 'Outdoors'
    | 'Sports'
    | 'Technology'
    | 'Fashion'
    | 'Food'
    | 'Photography'
    | 'Dance'
    | 'Volunteering'
  )[];
  religion:
    | 'Muslim'
    | 'Christian'
    | 'Jewish'
    | 'Hindu'
    | 'Buddhist'
    | 'Sikh'
    | 'Agnostic'
    | 'Atheist'
    | 'Other'
    | 'Prefer not to say';
  photos: PublicProfilePhoto[];
  prompts: PublicProfilePrompt[];
} | null;