// lib/api/generated/model/publicProfile.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { PublicDatingProfile } from './publicDatingProfile';

export interface PublicProfile {
  id: string;
  /** @nullable */
  chosenName: string | null;
  /** @nullable */
  avatarUrl: string | null;
  datingProfile: PublicDatingProfile | null;
}